+++ sync-for-payroll/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [typescript v0.2.0] sync-for-payroll
 ### Releases
 - [NPM v0.2.0] https://www.npmjs.com/package/@codat/sync-for-payroll/v/0.2.0 - sync-for-payroll
+
+## 2023-09-12 00:05:01
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payroll.yaml
+- Speakeasy CLI 1.82.5 (2.108.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.2.1] sync-for-payroll
+### Releases
+- [NPM v0.2.1] https://www.npmjs.com/package/@codat/sync-for-payroll/v/0.2.1 - sync-for-payroll